Adds unit testing
webdriver.io browser testing uses vite, so a number of changes were
made to make the build work under vite dev. Mostly making deps visible,
but fonts were tricky and the curren solution in `styles.ts` is a hack.

vite dev might be a suitable replacement for the existing devserver, but
some more work on the service worker is required to make vite build work
for prod. Need to configure vite export a separate binary.